src/container/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
opencv-python-headless==4.10.*
scenedetect[opencv-headless]==0.6.*
ffmpeg-python==0.2.*
numpy>=1.26,<2.0
pydantic>=2.0,<3.0
httpx>=0.27,<1.0
Pillow>=10.0,<11.0
